#135 - Refactor: 게시글 페이지 컴포넌트화
